Added -R option to randomize slideshow image order
When staring sxiv with -R command-line option and starting slideshow mode,
fileidx of the next image to be loaded will be randomized. This is done using
rand() and thus does not provide cryptographical-strength pseudo-random number
generation, but should nevertheless be good enough for viewing images in
random order.
This does *not* randomize the file order itself, e.g. when using shortcuts for
next/previous images; doing so would require either (a) randomizing the file
list itself on startup or (b) managing a linked list of the random sequence.
Both options have drawbacks in my point of view.