Do not crash hard
When creation of scaled image fails, the program shouldn't exit immediately. If it is changed to just report the error and use 1x1 random image instead, the program can continue showing other thumbnails. 

With this change I was able to inspect images in a directory of hundred of thousands files of various types.