Introduce a multiframe webp loader
Created a function `img_load_webp()` which does the same thing as the
function `img_load_gif()`, but for webp images. Added a small amount of
code to make sxiv call this function when the file given is a webp image
file (using file extension to check at first, then doing proper tests in
the function).